Make iteration args a generic void * struct element
This makes it easier to genericize the handling of iteration at the C level
without adding elements to the MMDB_tree_s struct for every possible iteration
scenario. The downside is that this is a fun way of doing an end run around
C's type checking. Wait, that sounds like a fun upside!