Remove some code that will never be reached during thaw for ::0/0 networks
When we freeze a tree with a ::0/0 network as its only data, the freeze code
ends up freezing two records, one for ::0/1 and one for 8000::/1, so we never
try to thaw ::0/0. This is just an artifact of how the tree is managed. Every
tree needs at least two records, so a tree with only one network has two
records with identical data.