Make removing reserved networks work by adding them as immutable empty records
This adds a new record type for this purpose: A fixed empty record.
These are like empty records in that they indicate there is no
information about the networks, but they are also immutable and you
cannot add networks under them (the latter being similar to alias
records).

We now add reserved networks with this record type. This means it will
not be possible to inadvertently add such networks at all when the
option to remove them is on. Previously that option worked by removing
such networks from the tree after the fact. This was problematic if you
used the tree prior to that point.